Don't use optionals for required connections
Prior to this commit it was possible for the code to "gracefully" handle
situations where a `DayStack` wasn't properly set up. Rather than
causing errors, I was careful to mark things as optional.

Fuck that.

If I make a mistake and don't have a `DayStack` setup with the required
subviews, then I want a runtime error.

This has the side effect of also making thing at the controller level
easier to work with so it's a win there too.