Add pace field to week view
This commit also changes behavior so that when you enter your in time,
but have not yet entered your out time, it'll show as a negative number,
which I hate, but I'm not sure what to do about it, so I'm going to just
live with it for now. What I probably want to do is somewhere set a
floor of zero, but I'm not sure where.