Add naive activity for updates
This is naive when there are two updates that overlap. In that case the
first one that finishes will turn off the activity indicator, leading
the user to believe that all network activity has completed.

A better approach would be to have the Router be the source of truth
here but that would be a total rethinking of how this all glues
together. In a world where the Router is the source of truth, you'd
probably use Notifications. The controller would listen and update the
activity indicator appropriately. The Router would then need an instance
to keep track of which tasks are in flight and which have finished. Only
when that queue was empty would it post the message that activity has
finished.

Another area of naivety here is the way the view animates. Its
currently possible for overlapping animations to occur because the
animations don't check to ensure that no animations are currently
happening. For this I would probably need another property on the view
to keep track of this and block one animation from starting while
another is finishing up.